Finish search styling
