+++ components/search/index.ts
@@ -1,4 +1,3 @@
 export * from "./BillSearch"
-export * from "./InstantSearch"
 export * from "./service"
 export * from "./TestimonySearch"
