+++ styles/bootstrap.scss
@@ -29,3 +29,7 @@ $headings-font-family: "Nunito", system-ui, -apple-system, "Segoe UI", Roboto,
 // Import fonts using the default bootstrap weights
 @import url("https://fonts.googleapis.com/css2?family=Archivo:wght@300;400;700&display=swap");
 @import url("https://fonts.googleapis.com/css2?family=Nunito:wght@300;400;700&display=swap");
+
+:root {
+  --bs-blue-100: #{$blue-100};
+}
